unify checkpointer and contextintegration interfaces
