+++ dd-java-agent/agent-profiling/profiling-controller-openjdk/src/main/java/com/datadog/profiling/controller/openjdk/JFREventContextIntegration.java
@@ -1,16 +1,30 @@
 package com.datadog.profiling.controller.openjdk;
 
+import com.datadog.profiling.controller.openjdk.events.EndpointEvent;
 import com.datadog.profiling.controller.openjdk.events.TimelineEvent;
 import com.datadog.profiling.utils.ExcludedVersions;
+import datadog.trace.api.EndpointTracker;
 import datadog.trace.api.Platform;
 import datadog.trace.api.Stateful;
+import datadog.trace.api.config.ProfilingConfig;
+import datadog.trace.bootstrap.config.provider.ConfigProvider;
+import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import datadog.trace.bootstrap.instrumentation.api.ProfilerContext;
 import datadog.trace.bootstrap.instrumentation.api.ProfilingContextIntegration;
+import jdk.jfr.EventType;
 
 public class JFREventContextIntegration implements ProfilingContextIntegration {
 
   public JFREventContextIntegration() {
     ExcludedVersions.checkVersionExclusion();
+    // Note: Loading CheckpointEvent when JFRCheckpointer is loaded is important because it also
+    // loads JFR classes - which may not be present on some JVMs
+    EventType.getEventType(EndpointEvent.class);
+    isEndpointCollectionEnabled =
+        ConfigProvider.getInstance()
+            .getBoolean(
+                ProfilingConfig.PROFILING_ENDPOINT_COLLECTION_ENABLED,
+                ProfilingConfig.PROFILING_ENDPOINT_COLLECTION_ENABLED_DEFAULT);
   }
 
   // native image process will enable context integration immediately - the value will get
@@ -18,6 +32,7 @@ public class JFREventContextIntegration implements ProfilingContextIntegration {
   // java agent process will enable context integration when the agent is started and onStart is
   // called
   private volatile boolean isStarted = !Platform.isNativeImageBuilder();
+  private final boolean isEndpointCollectionEnabled;
 
   @Override
   public void onStart() {
@@ -39,5 +54,19 @@ public class JFREventContextIntegration implements ProfilingContextIntegration {
   @Override
   public String name() {
     return "jfr";
+  }
+
+  @Override
+  public void onRootSpanFinished(AgentSpan rootSpan, EndpointTracker tracker) {
+    if (isEndpointCollectionEnabled && tracker != null) {
+      tracker.endpointWritten(rootSpan);
+    }
+  }
+
+  @Override
+  public EndpointTracker onRootSpanStarted(AgentSpan rootSpan) {
+    return isEndpointCollectionEnabled
+        ? new EndpointEvent(rootSpan.getSpanId())
+        : EndpointTracker.NO_OP;
   }
 }
