--- dd-java-agent/testing/src/main/groovy/datadog/trace/agent/test/checkpoints/TestEndpointCheckpointer.groovy
@@ -1,18 +1,0 @@
-package datadog.trace.agent.test.checkpoints
-
-import datadog.trace.api.EndpointCheckpointer
-import datadog.trace.api.EndpointTracker
-import datadog.trace.bootstrap.instrumentation.api.AgentSpan
-
-class TestEndpointCheckpointer implements EndpointCheckpointer {
-
-
-  @Override
-  void onRootSpanFinished(AgentSpan rootSpan, EndpointTracker tracker) {
-  }
-
-  @Override
-  EndpointTracker onRootSpanStarted(AgentSpan rootSpan) {
-    return EndpointTracker.NO_OP
-  }
-}
